feat and fix : fixing landing page navbar and banner, create paket page
