--- src/app/(auth)/daftar/page.tsx
@@ -1,15 +1,0 @@
-"use client";
-import Navbar from "../../home-page/_components/Navbar";
-import React from "react";
-import RegisterSection from "./_components/Register";
-
-const Daftar = () => {
-  return (
-    <>
-      <Navbar />
-      <RegisterSection />
-    </>
-  );
-};
-
-export default Daftar;
