--- src/app/(auth)/forgot-pass/page.tsx
@@ -1,15 +1,0 @@
-"use client";
-import Navbar from "../../home-page/_components/Navbar";
-import React from "react";
-import ForgotPassSection from "./_components/ForgotPass";
-
-const Forgot: React.FC = () => {
-  return (
-    <>
-      <Navbar />
-      <ForgotPassSection />
-    </>
-  );
-};
-
-export default Forgot;
