--- src/app/(auth)/login/page.tsx
@@ -1,15 +1,0 @@
-"use client";
-import Navbar from "../../home-page/_components/Navbar";
-import React from "react";
-import LoginSection from "./_components/Login";
-
-const Login: React.FC = () => {
-  return (
-    <>
-      <Navbar />
-      <LoginSection />
-    </>
-  );
-};
-
-export default Login;
