--- src/app/(auth)/re-type-pass/page.tsx
@@ -1,15 +1,0 @@
-"use client";
-import Navbar from "../../home-page/_components/Navbar";
-import React from "react";
-import ReTypePassSection from "./_components/ReTypePass";
-
-const ReTypePass: React.FC = () => {
-  return (
-    <>
-      <Navbar />
-      <ReTypePassSection />
-    </>
-  );
-};
-
-export default ReTypePass;
