+++ src/app/(landing)/(auth)/daftar/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+import React from "react";
+import RegisterSection from "./_components/Register";
+
+const Daftar = () => {
+  return (
+    <>
+      <RegisterSection />
+    </>
+  );
+};
+
+export default Daftar;
