+++ src/app/(landing)/(auth)/forgot-pass/_components/ForgotPass/index.tsx
