+++ src/app/(landing)/(auth)/forgot-pass/page.tsx
@@ -1,0 +1,15 @@
+"use client";
+import Navbar from "../../../../components/Navbar";
+import React from "react";
+import ForgotPassSection from "./_components/ForgotPass";
+
+const Forgot: React.FC = () => {
+  return (
+    <>
+      <Navbar />
+      <ForgotPassSection />
+    </>
+  );
+};
+
+export default Forgot;
