+++ src/app/(landing)/(auth)/login/page.tsx
@@ -1,0 +1,14 @@
+"use client";
+import Navbar from "../../../../components/Navbar";
+import React from "react";
+import LoginSection from "./_components/Login";
+
+const Login: React.FC = () => {
+  return (
+    <>
+      <LoginSection />
+    </>
+  );
+};
+
+export default Login;
