+++ src/app/(landing)/(auth)/re-type-pass/_components/ReTypePass/index.tsx
