+++ src/app/(landing)/(auth)/re-type-pass/page.tsx
@@ -1,0 +1,15 @@
+"use client";
+import Navbar from "../../../../components/Navbar";
+import React from "react";
+import ReTypePassSection from "./_components/ReTypePass";
+
+const ReTypePass: React.FC = () => {
+  return (
+    <>
+      <Navbar />
+      <ReTypePassSection />
+    </>
+  );
+};
+
+export default ReTypePass;
