+++ src/app/(landing)/home-page/_components/FaqItem.tsx
